@@ -12,8 +12,8 @@ import {LiquidExtension} from './LiquidExtension';
 import {AppManifest} from './types';
 import * as manifestSchema from './types/AppManifest.schema.json';
 import {SchemaObjects, SchemaObject} from './types/SchemaObject';
-import deepFreeze = require('deep-freeze');
-import glob = require('glob');
+import * as deepFreeze from 'deep-freeze';
+import * as glob from 'glob';
 
 interface SerializedRuntime {
   appManifest: AppManifest;
@@ -56,7 +56,7 @@ export class Runtime {
     return this.dirName;
   }
 
-  // eslint-disable-next-line @typescript-eslint/ban-types
+  // eslint-disable-next-line @typescript-eslint/no-restricted-types
   public async getFunctionClass<T extends Function>(name: string): Promise<new (request: Request) => T> {
     const functions = this.manifest.functions;
     if (!functions || !functions[name]) {
