@@ -160,7 +160,7 @@ describe('validateFunctions', () => {
 
   it('detects global function defining a installation resolution', async () => {
     const runtime = Runtime.fromJson(JSON.stringify({appManifest, dirName: '/tmp/foo'}));
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+
     runtime.manifest.functions!.global_foo.installation_resolution = {type: 'HEADER', key: 'foo'};
     const getFunctionClass = jest.spyOn(Runtime.prototype, 'getFunctionClass')
       .mockImplementation((name) => Promise.resolve(name === 'foo' ? ProperFoo : ProperGlobalFoo));
@@ -172,7 +172,7 @@ describe('validateFunctions', () => {
 
   it('detects a invalid JSONPath expression', async () => {
     const runtime = Runtime.fromJson(JSON.stringify({appManifest, dirName: '/tmp/foo'}));
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+
     runtime.manifest.functions!.foo.installation_resolution = {
       type: 'JSON_BODY_FIELD', key: '/test/foo'
     };
