@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
+
 import * as deepFreeze from 'deep-freeze';
 import 'jest';
 import {Runtime} from '../../Runtime';
