@@ -32,7 +32,7 @@ describe('jobs', () => {
 
   describe('async local store configured', () => {
     it('uses local job Api if not configured', async () => {
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+
       const getEndpointsFn = jest.spyOn(LocalJobApi.prototype, 'trigger');
 
       expect(() => jobs.trigger('foot', {})).toThrow();
